util/mon: use int32 instead of int64 for two things
This commit switches int64 to int32 for `BoundAccount.reserved` and
`BytesMonitor.poolAllocationSize`. The main reason for doing this is the
desire of introducing a new boolean to `BoundAccount` without increasing
the struct's memory footprint. These two fields track the buffer that
the account keeps in order to avoid going to the monitor on every `Grow`
call (i.e. it exists to optimize `Grow -> Shrink -> Grow -> Shrink ...`
sequence by serving it without accessing the monitor, which would
require locking). Generally, this buffer is very small (on the order of
10-100KiB) which easily fits into int32 range.

The only use case when the reserved buffer can be large is when we use
the earmarked account. It is currently used in one place (by the bulk
adder), and there we use values that are 32MiB by default, so they still
easily fit into the int32 range. Some of these reservations are
configurable via cluster settings, so this commit introduces validation
that those settings are not set to a value larger than MaxInt32. Note
that if some cluster has already set the settings to a value larger than
2GiB, then when the node is upgraded to the binary that includes this
fix the setting will be reset to its default value. This seems
acceptable.

Release note: None

+++ pkg/sql/colexec/external_distinct_test.go
@@ -139,8 +139,8 @@ func TestExternalDistinctSpilling(t *testing.T) {
 	// spilling occurs.
 	nBatchesOutputByInMemoryOp := 2 + rng.Int63n(2)
 	memoryLimitBytes := nBatchesOutputByInMemoryOp * batchMemEstimate
-	if memoryLimitBytes < mon.DefaultPoolAllocationSize {
-		memoryLimitBytes = mon.DefaultPoolAllocationSize
+	if memoryLimitBytes < int64(mon.DefaultPoolAllocationSize) {
+		memoryLimitBytes = int64(mon.DefaultPoolAllocationSize)
 		nBatchesOutputByInMemoryOp = memoryLimitBytes / batchMemEstimate
 	}
 	flowCtx.Cfg.TestingKnobs.MemoryLimitBytes = memoryLimitBytes
